Inlining now respects preexisting style attributes
Previously in a tag like this:
    <div style="color: red;">

InlineStyler would push those rules up to the front and apply
all its CSS rules after, even though in CSS anything in a
`style` attribute should be the last thing applied. Like this:

    <div style="color: red; color: blue">

when it should do this:

    <div style="color: blue; color: red;">

This commit changes that by placing a "bookmark" between
previously-existing styles and those being added by the
inliner. That way anything that already existed gets applied
last, in expected CSS order.