Improved selector score-sorting rules
If you had two properties where one overrides the other, for example:

    font-size: 30px;
    font-size: 100%;

And both properties came from a selector with equal specificity (both having a score of 0-0-0-1, for example), sorting would still place "font-size: 30px" after "font-size: 100%" because, alphanumerically, "30px" comes after "100%".

This change accounts for a property's placement in a list of properties to prevent that from happening.